dev/4_microservices/random/random-lang/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4
google-cloud-bigquery ~= 3.0.1